[Refactor] FileUploadController, CloudFileUploadService 패키지 구조 변경 (#41)
