+++ backend/src/main/java/org/example/backend/File/Service/CloudFileUploadService.java
@@ -1,4 +1,4 @@
-package org.example.backend.File;
+package org.example.backend.File.Service;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3Client;
